[WOUNDMED] Bleeding Refactor Part 1 - HealingSystem.cs
