+++ Content.Shared/Medical/_Goobstation/HealingSystemEvent.cs
@@ -1,0 +1,14 @@
+using Content.Shared.Damage;
+using Content.Shared.DoAfter;
+using Content.Shared.Medical.Healing;
+
+namespace Content.Shared.Medical._Goobstation;
+
+/// <summary>
+/// Event raised when a wound is being medicated.
+/// </summary>
+[ByRefEvent]
+public record struct HealingSystemEvent(EntityUid Entity)
+{
+    public bool AnythingToDo = false;
+}
